chore: update comsp
