+++ apps/embed/src/components/Video.tsx
@@ -11,40 +11,10 @@ import truncate from 'utils/functions/truncate'
 import VideoPlayer from 'web-ui/VideoPlayer'
 
 import MetaTags from './MetaTags'
-import VideoOverlay from './VideoOverlay'
+import TopOverlay from './TopOverlay'
 
 type Props = {
   video: Publication
-}
-
-type OverlayProps = {
-  playerRef: HTMLMediaElement | undefined
-  video: Publication
-  clicked: boolean
-}
-
-const TopOverlay: FC<OverlayProps> = ({ playerRef, video, clicked }) => {
-  const [showVideoOverlay, setShowVideoOverlay] = useState(true)
-
-  useEffect(() => {
-    if (!playerRef) return
-    playerRef.onpause = () => {
-      setShowVideoOverlay(true)
-    }
-    playerRef.onplay = () => {
-      setShowVideoOverlay(false)
-    }
-  }, [playerRef])
-
-  return (
-    <div
-      className={`${
-        showVideoOverlay ? 'visible' : 'invisible'
-      } transition-all duration-200 ease-in-out group-hover:visible`}
-    >
-      <VideoOverlay video={video} clicked={clicked} />
-    </div>
-  )
 }
 
 const Video: FC<Props> = ({ video }) => {
@@ -83,50 +53,48 @@ const Video: FC<Props> = ({ video }) => {
   }
 
   return (
-    <div className="w-screen h-screen">
+    <div className="relative group w-screen h-screen">
       <MetaTags
         title={truncate(video?.metadata?.name as string, 60)}
         description={truncate(video?.metadata?.description as string, 100)}
         image={thumbnailUrl}
         videoUrl={getPublicationMediaUrl(video)}
       />
-      <div className="relative group">
-        {clicked ? (
-          <VideoPlayer
-            refCallback={refCallback}
-            permanentUrl={getPublicationMediaUrl(video)}
-            posterUrl={thumbnailUrl}
-            publicationId={video.id}
-            currentTime={currentTime}
-            options={{ autoPlay: isAutoPlay, muted: isAutoPlay, loop: isLoop }}
+      {clicked ? (
+        <VideoPlayer
+          refCallback={refCallback}
+          permanentUrl={getPublicationMediaUrl(video)}
+          posterUrl={thumbnailUrl}
+          publicationId={video.id}
+          currentTime={currentTime}
+          options={{ autoPlay: isAutoPlay, muted: isAutoPlay, loop: isLoop }}
+        />
+      ) : (
+        <div
+          className="grid place-items-center"
+          onClick={onClickOverlay}
+          role="button"
+        >
+          <img
+            src={thumbnailUrl}
+            className="h-full w-full"
+            alt={video.metadata.name ?? video.profile.handle}
+            draggable={false}
           />
-        ) : (
-          <div
-            className="grid place-items-center"
-            onClick={onClickOverlay}
-            role="button"
-          >
+          <button className="xl:p-5 p-3 rounded-full bg-indigo-500 absolute">
             <img
-              src={thumbnailUrl}
-              className="h-full w-full"
-              alt={video.metadata.name ?? video.profile.handle}
+              className="w-5 h-5"
+              src={imageCdn(
+                `${STATIC_ASSETS}/images/brand/white.svg`,
+                'avatar'
+              )}
+              alt="play"
               draggable={false}
             />
-            <button className="xl:p-5 p-3 rounded-full bg-indigo-500 absolute">
-              <img
-                className="w-5 h-5"
-                src={imageCdn(
-                  `${STATIC_ASSETS}/images/brand/white.svg`,
-                  'avatar'
-                )}
-                alt="play"
-                draggable={false}
-              />
-            </button>
-          </div>
-        )}
-        <TopOverlay playerRef={playerRef} video={video} clicked={clicked} />
-      </div>
+          </button>
+        </div>
+      )}
+      <TopOverlay playerRef={playerRef} video={video} clicked={clicked} />
     </div>
   )
 }
